Remove bin and obj, fixed exercise1

+++ exercise1/exercise1.cs
@@ -4,11 +4,11 @@ class Program1
 {
     static void Main()
     {
-        int degreesСelsius = 0;
+        double degreesСelsius = 0;
         Console.Write("Введите температуру в градусах Цельсия: ");
         string? input = Console.ReadLine();
 
-        if (input != null && int.TryParse(input, out degreesСelsius))
+        if (input != null && double.TryParse(input, out degreesСelsius))
         {
             Console.Write($"{degreesСelsius}°C = { degreesСelsius * 9 / 5 + 32}°F");
         }
